weather_app without env/requirements.txt
flask==2.2.3
psycopg2-binary==2.9.5
flasgger==0.9.5
flake8==6.0.0
flask-sqlalchemy==3.0.3
